ensure lock release after stream/range process exit (bz://504)
Instead of waiting for a gen_server:call to signal the end of a stream or range process,
we just catch the EXIT message from the link and release the locks that the process was
holding then.

If the process exits abnormally, this code covers sending the '$end_of_table' message.  It
might be useful to send some other variety of message, but since none of the callers of
:stream are prepared to do anything about a mid-stream failure, it's probably just as well
to declare the end of the stream at the failure point.