move buffer converter process to a supervisor-monitored gen_server (bz://504)
It was possible for the buffer converter to get into a fast-death restart spin.  I've
moved it out to a gen_server monitored by a supervisor that will die if the process has to
be restarted more than twice in one second.  mi_server monitors the supervisor, and exits
if that max restart frequency is reached.