bz://727 Part 1: Broken homedir for riak user
Fixed the adduser call to place the homedir in /opt/riaksearch for the riak
user but only when the user doesn't already exist.  This should prevent use
from breaking any existing installations.

Also fix a copy'n'paste error in the r.preserve script that thought it was
part of the riak-ee package instead of the riak-search package.