Filter node annotations
Commit d3ccf9 in riak_core added annotations to the preflist which
vnode_master command doesn't take account for.  For now the clients
will just have to filter out the annotations themselves but we should
look at making the preflist an opaque type to avoid stuff like this in
the future.