Implement handle_exit
Since 14.0 Riak Core has added an "optional" handle_exit callback for
vnodes to implement when they want to handle EXIT signals from
processes linked to the vnode container.  However, it doesn't special
case a normal exit and therefore the search vnode has to implement
this callback to handle that case.

The handle_exit callback simply ignores any exit signals by returning
noreply, this is consistent with it's behavior in 0.14.0.