Add ability to run static query with bench driver
I wanted the ability to bench a static query against a prepopulated
Search cluster.  This is now possible by binding a list to the
`riaksearch_query` config.  It should be a list of 2-3 args

[Index, Query, Filter]

e.g. ["tweets", "created_at:[20100113T032200 TO 20100113T032500]",
      "text":"earthquake"]

Furthermore, you can also set `riaksearch_expected` to an integer
and it will check to make sure the correct number of results were
returned.