Remove JS extractor support
bz1144

It's been decided that JavaScript extractor support will be removed.
The main reasons for doing so are:

1. This feature has never worked so no existing users are being broken.

2. The documentation around this feature isn't entirely clear and the
code to support it is ugly, IMO.

3. Like any other JS support in Riak it has to pay the penalty of
shipping bytes between the Erlang and JS VMs.  There is also potential
contention for JS VMs depending on the size of the pool and load on
the system.

4. Finally, the main reason I can think of for anyone to use the JS
extractor support is to pull apart incoming JSON objects.  This is
already handled via an Erlang extractor that is invoked any time it
sees an `application/json` content-type header.  This will be much
more efficient than using a JS extractor.