Fix search-doc cmd
az504

The search-doc command was accessing the `#riak_idx_doc` record
directly which caused a regression to slip in because at some
point its internal format changed.  The `riak_indexed_doc` module
is meant to be used as an accessor to this structure so I simply
changed search-doc to use that.

I also cleaned up the usage statement for search-cmd.