Load 1.0 code binary during 0.14.2->1.0.0 rolling upgrade
In 14.2 we ship a closure across the wire that is created in the
`riak_search_client` module.  Lambdas are shipped with a MD5 of the entire
module's code and if the MD5s of the module on both sides don't agree then
a `badfun` will be thrown when trying to execute the lambda.

In the case of 14.2->1.0.0, during the intermediate period where there is
a mixed cluster, if you have a query that is executed on a different node
than where the query was issued and the two nodes have different versions of
Riak then a `badfun` will ensue and the query will timeout.  This is because
there where code changes in `riak_search_client` which caused the MD5 to
change.

This patch addresses this by detecting a rolling upgrade from 0.14.2->1.0.0
and loading the 1.0.0 version of `riak_search_client` across all nodes in the
cluster.  This is a necessary evil to prevent queries from failing during
rolling upgrade.  In the future we should make a fix to  just ship the data
instead of a closure. Both pieces of data, the schema and filter ops could be
created on the other node.  If we do need to ship a fun it should be limited
to export funs.