Clear schema cache after schema update
After updating the schema make sure to clear the cache on all nodes to
make sure the update propagates to all nodes immediately.  If one of
the remote calls fails then that node will still use the old schema so
it's not perfect.  In the future the schema should be gossiped in a
manner similar to the ring in order to get convergence.