Remove some remaining qilr stuff
There were still bits of qilr clinging to the code base.  I left the
qilr stuff in the one EQC test alone for now as perhaps that test
could be resurrected in the future or used as inspiration.  In any
case if that test goes the entire module should be deleted at once
rather than just removing some of it.