Fix search tests
Host/port was changed from being a riak_core config to a riak_api config
a little while ago.
After the default setting for allow_mult (return siblings) was changed,
search has been unhappy as it relied on it. This change enforces the old
default in search buckets to allow it to exist in this brave new world