Allow search to start if security is enabled, just disable its APIs
It turns out that simply not starting search if security is enabled did
not play well with the new 'riak-admin security enable' command, because
it was decoupled from node startup. Instead, simply disable the endpoint
APIs that search exposes if security is enabled, making it more friendly
to runtime changes.